ESRGAN_Streamlit_app/requirements.txt
streamlit==1.39.0
torch==2.0.1+cpu
torchvision==0.15.2+cpu
Pillow==7.1.2
numpy==1.26.4
opencv-python==4.10.0
scikit-image==0.24.0
